+++ lib/features/classroom/presentation/pages/classroom_page.dart
@@ -1,15 +1,14 @@
 import 'package:flutter/material.dart';
 
 import '../../../../core/widgets/main_app_bar.dart';
+import '../widgets/classroom_start_row.dart';
 import '../widgets/first_part_of_row_widget.dart';
-import '../widgets/couse_start_row.dart';
-import '../widgets/second_part_of_row_widget.dart';
 
-class CoursePage extends StatelessWidget {
+class ClassroomPage extends StatelessWidget {
   /// The page route name.
-  static const routeName = 'course';
+  static const routeName = 'classroom';
 
-  const CoursePage({super.key});
+  const ClassroomPage({super.key});
 
   @override
   Widget build(BuildContext context) {
